// src/app/models/github-users-data.ts
import { Data } from '@angular/router';

export interface IGithubUsersData {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id: number;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  site_admin: boolean;
  name: string;
  location: string;
  followers: number;
  following: number;
  created_at: string;
  updated_at: string;
}